requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
